feat(datasets): make  default UI tab in datasets detail page
