+++ web/src/features/navigation/utils/dataset-tabs.ts
@@ -1,6 +1,6 @@
 export const DATASET_TABS = {
-  RUNS: "runs",
   ITEMS: "items",
+  EXPERIMENTS: "experiments",
 } as const;
 
 export type DatasetTab = (typeof DATASET_TABS)[keyof typeof DATASET_TABS];
@@ -8,14 +8,14 @@ export type DatasetTab = (typeof DATASET_TABS)[keyof typeof DATASET_TABS];
 export const getDatasetTabs = (projectId: string, datasetId: string) => {
   return [
     {
-      value: DATASET_TABS.RUNS,
-      label: "Experiments",
-      href: `/project/${projectId}/datasets/${datasetId}`,
-    },
-    {
       value: DATASET_TABS.ITEMS,
       label: "Items",
       href: `/project/${projectId}/datasets/${datasetId}/items`,
+    },
+    {
+      value: DATASET_TABS.EXPERIMENTS,
+      label: "Experiments",
+      href: `/project/${projectId}/datasets/${datasetId}/experiments`,
     },
   ];
 };
